Rename `graph::test_types` to `test_utils`
Also move the code for the module from `graph.rs` to
`graph/test_utils`.

This is to make room for a component graph builder for tests, that
will be added in the next commit.
