+++ src/graph/meter_roles.rs
@@ -134,7 +134,7 @@ mod tests {
     use crate::component_category::BatteryType;
     use crate::component_category::EvChargerType;
     use crate::error::Error;
-    use crate::graph::test_types::{TestComponent, TestConnection};
+    use crate::graph::test_utils::{TestComponent, TestConnection};
     use crate::ComponentCategory;
     use crate::InverterType;
