+++ src/graph/validation/validate_neighbors.rs
@@ -141,7 +141,7 @@ mod tests {
     use super::*;
     use crate::component_category::BatteryType;
     use crate::component_category::EvChargerType;
-    use crate::graph::test_types::{TestComponent, TestConnection};
+    use crate::graph::test_utils::{TestComponent, TestConnection};
     use crate::ComponentCategory;
     use crate::ComponentGraph;
     use crate::InverterType;
